# Manage the Tracker's lifetime with `shared_ptr`