Whenever any hook needs to make use of the Tracker singleton, have it atomically grab a reference to a `shared_ptr` singleton. When stopping tracking, unset that singleton, and then wait for all of the remaining references to it to be destroyed before really destroying the Tracker.