@pablogsal 2023-01-12
I did some benchmarking and this solution is 60% slower than main :(:

Main:
```
hyperfine 'memray run --trace-python-allocators -o /dev/null -m test test_list test_list test_list test_list'
Benchmark 1: memray run --trace-python-allocators -o /dev/null -m test test_list test_list test_list test_list
  Time (mean ± σ):      1.657 s ±  0.021 s    [User: 1.598 s, System: 0.062 s]
  Range (min … max):    1.640 s …  1.709 s    10 runs
```	

This PR:
```
hyperfine 'memray run --trace-python-allocators -o /dev/null -m test test_list test_list test_list test_list'
Benchmark 1: memray run --trace-python-allocators -o /dev/null -m test test_list test_list test_list test_list
  Time (mean ± σ):      2.748 s ±  0.029 s    [User: 2.682 s, System: 0.072 s]
  Range (min … max):    2.715 s …  2.806 s    10 runs
```