@godlygeek 2023-01-12
Well, that's very sad. This looks very elegant, but obviously we can't accept that kind of overhead. It looks like it's coming from locking in the `shared_ptr` implementation. That's supposed to be cheap, but when you do something on every malloc or free, the cost adds up.

Oh well.